requirements.txt
# Dependências principais
fastapi==0.111.0
uvicorn==0.30.0
SQLAlchemy==2.0.29
pydantic==2.8.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.0

# Dependências de desenvolvimento
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
ruff==0.1.6
black==23.11.0

# Base de dados
psycopg2-binary==2.9.9  # Para PostgreSQL em produção
